requirements.txt
streamlit
google-generativeai
python-docx
docx2txt
PyPDF2
python-dotenv
